Improve command execution handling
Currently we load shell from the shell env, this is a bit surprising so we can instead use
sh. This lets us do everything we want, and is always going to exist and
be set. This also change the error message to be clearer that it's an
exec that's failed