fix: Switch to clipboard libary that is maintained
This switches to a more readilly updated library. The current library
now has two vulnerabilities that need to be looked after, so I am no
longer comfortable using it. This new one has one vulnerability, but
that vulnerability is new, and they aren't using the effected function.

Clipboard libraries are not in a good place in rust.